# refactor(vacuum statuses): reduce amount of db queries and load for each query - improve performance