--- a/app/lib/vacuum/statuses_vacuum.rb
+++ b/app/lib/vacuum/statuses_vacuum.rb
@@ -14,33 +14,34 @@ class Vacuum::StatusesVacuum
   private
 
   def vacuum_statuses!
-    statuses_scope.find_in_batches do |statuses|
+    statuses_scope.in_batches do |statuses|
       # Side-effects not covered by foreign keys, such
       # as the search index, must be handled first.
+      statuses.direct_visibility
+              .includes(mentions: :account)
+              .find_each do |status|
+        # TODO: replace temporary solution - call of private model method
+        status.send(:unlink_from_conversations)
+      end
+      remove_from_search_index(statuses.ids) if Chewy.enabled?
 
-      remove_from_account_conversations(statuses)
-      remove_from_search_index(statuses)
-
-      # Foreign keys take care of most associated records
-      # for us. Media attachments will be orphaned.
-
-      Status.where(id: statuses.map(&:id)).delete_all
+      # Foreign keys take care of most associated records for us.
+      # Media attachments will be orphaned.
+      statuses.delete_all
     end
   end
 
   def statuses_scope
-    Status.unscoped.kept.where(account: Account.remote).where(Status.arel_table[:id].lt(retention_period_as_id)).select(:id, :visibility)
+    Status.unscoped.kept
+          .joins(:account).merge(Account.remote)
+          .where('statuses.id < ?', retention_period_as_id)
   end
 
   def retention_period_as_id
     Mastodon::Snowflake.id_at(@retention_period.ago, with_random: false)
   end
 
-  def remove_from_account_conversations(statuses)
-    Status.where(id: statuses.select(&:direct_visibility?).map(&:id)).includes(:account, mentions: :account).each(&:unlink_from_conversations)
-  end
-
-  def remove_from_search_index(statuses)
-    with_redis { |redis| redis.sadd('chewy:queue:StatusesIndex', statuses.map(&:id)) } if Chewy.enabled?
+  def remove_from_search_index(status_ids)
+    with_redis { |redis| redis.sadd('chewy:queue:StatusesIndex', status_ids) }
   end
 end