This PR is the result of a PR discussion:
https://github.com/mastodon/mastodon/pull/21297#discussion_r1028234021

- [refactor(statuses_vacuum): improve query performance](https://github.com/mastodon/mastodon/pull/21487/commits/b0974f7a07a76cd606c6a0b1d9d8b17fd31b30f9)   
  - fix statuses_scope having sub-select for Account.remote scope by
  `joins(:account).merge(Account.remote)`
  - fix statuses_scope unnecessary use of `Status.arel_table[:id].lt`
  because it is inexplicit, bad practice and even slower than normal
  `.where('statuses.id < ?'`
  - fix statuses_scope remove select(:id, :visibility) for having reusable
  active record query batches (no re queries)
  - fix vacuum_statuses! to use in_batches instead of find_in_batches,
  because in_batches delivers a full-blown active record query result,
  instead of an array — no re-queries necessary
  - send(:unlink_from_conversations) not to perform another db query, but
  reuse the in_batches result instead.
  - remove now obsolete remove_from_account_conversations method
  - remove_from_search_index uses array of ids, instead of mapping
  the ids from an array – this should be more efficient
  - use the in_batches scope to call delete_all, instead of running
  another db query for this – because it is again more efficient
  - add To-do comment for calling model's private method with send

## Additionally

- [refactor(statuses_vacuum): remove dead code - unused](https://github.com/mastodon/mastodon/pull/21487/commits/ab9afb67e16237148168ee12c5f93deb93609e79)
- [refactor(statuses_vacuum): make retention_period present test explicit](https://github.com/mastodon/mastodon/pull/21487/commits/d06892f91db0a99e2409132ae108f102daab47c6)
- [refactor(status): simplify unlink_from_conversations](https://github.com/mastodon/mastodon/pull/21487/commits/8826e931d68089a8b24a97fa59fd4e1904663591)
  - add has_many through relation mentioned_accounts
  - use model scope local instead of method call `Status#local?`
  - more readable add account to inbox_owners when account.local?

[refactor(status): searchable_by way less sub selects](https://github.com/mastodon/mastodon/pull/21487/commits/06c0b8503619203c34751e7f8df683dd45ae4c1d)  
These queries all included a sub-select. Doing the same with a joins should be more efficient.  
Since this method does 5 such queries, this should be significant, since it technically halves the query count.   

This is how it was:   
```ruby 
[3] pry(main)> Status.first.mentions.where(account: Account.local, silent: false).explain
  Status Load (1.6ms)  SELECT "statuses".* FROM "statuses" WHERE "statuses"."deleted_at" IS NULL ORDER BY "statuses"."id" DESC LIMIT $1  [["LIMIT", 1]]
  Mention Load (1.5ms)  SELECT "mentions".* FROM "mentions" WHERE "mentions"."status_id" = $1 AND "mentions"."account_id" IN (SELECT "accounts"."id" FROM "accounts" WHERE "accounts"."domain" IS NULL) AND "mentions"."silent" = $2  [["status_id", 109382923142288414], ["silent", false]]
=> EXPLAIN for: SELECT "mentions".* FROM "mentions" WHERE "mentions"."status_id" = $1 AND "mentions"."account_id" IN (SELECT "accounts"."id" FROM "accounts" WHERE "accounts"."domain" IS NULL) AND "mentions"."silent" = $2 [["status_id", 109382923142288414], ["silent", false]]
                                                    QUERY PLAN
------------------------------------------------------------------------------------------------------------------
 Nested Loop  (cost=0.15..23.08 rows=1 width=41)
   ->  Seq Scan on accounts  (cost=0.00..10.90 rows=1 width=8)
         Filter: (domain IS NULL)
   ->  Index Scan using index_mentions_on_account_id_and_status_id on mentions  (cost=0.15..8.17 rows=1 width=41)
         Index Cond: ((account_id = accounts.id) AND (status_id = '109382923142288414'::bigint))
         Filter: (NOT silent)
(6 rows)
```
  This is how it is with this change:   
```ruby
[4] pry(main)> Status.first.mentions.joins(:account).merge(Account.local).active.explain
  Status Load (1.7ms)  SELECT "statuses".* FROM "statuses" WHERE "statuses"."deleted_at" IS NULL ORDER BY "statuses"."id" DESC LIMIT $1  [["LIMIT", 1]]
  Mention Load (0.7ms)  SELECT "mentions".* FROM "mentions" INNER JOIN "accounts" ON "accounts"."id" = "mentions"."account_id" WHERE "mentions"."status_id" = $1 AND "accounts"."domain" IS NULL AND "mentions"."silent" = $2  [["status_id", 109382923142288414], ["silent", false]]
=> EXPLAIN for: SELECT "mentions".* FROM "mentions" INNER JOIN "accounts" ON "accounts"."id" = "mentions"."account_id" WHERE "mentions"."status_id" = $1 AND "accounts"."domain" IS NULL AND "mentions"."silent" = $2 [["status_id", 109382923142288414], ["silent", false]]
                                                    QUERY PLAN
------------------------------------------------------------------------------------------------------------------
 Nested Loop  (cost=0.15..23.08 rows=1 width=41)
   ->  Seq Scan on accounts  (cost=0.00..10.90 rows=1 width=8)
         Filter: (domain IS NULL)
   ->  Index Scan using index_mentions_on_account_id_and_status_id on mentions  (cost=0.15..8.17 rows=1 width=41)
         Index Cond: ((account_id = accounts.id) AND (status_id = '109382923142288414'::bigint))
         Filter: (NOT silent)
(6 rows)
```