@ClearlyClaire 2022-11-24
> * fix statuses_scope unnecessary use of `Status.arel_table[:id].lt`
>       because it is inexplicit, bad practice and even slower than normal
>       `.where('statuses.id < ?'`

I understand the other arguments, but not the performance one. Out of personal curiosity, why would it be slower?