@casaper 2022-11-24
> > * fix statuses_scope unnecessary use of `Status.arel_table[:id].lt`
> >       because it is inexplicit, bad practice and even slower than normal
> >       `.where('statuses.id < ?'`
> 
> I understand the other arguments, but not the performance one. Out of personal curiosity, why would it be slower?

I Honestly don't know.
But the query analysis was always turning out a few 1/10 mil slower, when zhe rest of the query was the same.

So I wonder just as much.  
Assuming the query analysis is not wrong somehow, I assume it is slower.