@ClearlyClaire 2022-11-24
> I Honestly don't know. But the query analysis was always turning out a few 1/10 mil slower, when zhe rest of the query was the same.
> 
> So I wonder just as much. Assuming the query analysis is not wrong somehow, I assume it is slower.

I'd say it's probably a fluke since the SQL query should be the same :thinking:

Similarly, on your example in the PR description, the query planner gives the exact same result for the two queries, so the old query may have been just as efficient in practice (for your test database at least). That being said, the new query looks more idiomatic, and the PR as a whole fixes all sorts of other issues, so it definitely looks good to me.